Convert snippet to a component
